% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMLR_noc.R
\name{CIMLR_noc}
\alias{CIMLR_noc}
\title{CIMLR_noc}
\usage{
CIMLR_noc(
  X,
  no.dim = NA,
  k = 10,
  if.impute = FALSE,
  normalize = FALSE,
  cores.ratio = 1
)
}
\arguments{
\item{X}{a list of multi-omic data each of which is an (m x n) data matrix of measurements of cancer patients}

\item{no.dim}{number of dimensions}

\item{k}{tuning parameter}

\item{cores.ratio}{ratio of the number of cores to be used when computing the multi-kernel}
}
\value{
similarities of each omics data

list of 5 elements describing the clusters obtained by CIMLR_noc, of which are the resulting clusters:
S = similarities computed by CIMLR,
F = results from network diffiusion,
alphaK = clustering coefficients,
execution.time = execution time of the present run,
converge = iterative convergence values by T-SNE,

similarities S computed by CIMLR
}
\description{
perform the CIMLR clustering algorithm
}
\examples{
CIMLR_noc(X = exampledata$mRNA, cores.ratio = 0)

}
