% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.kernels2.R
\name{combine.kernels2}
\alias{combine.kernels2}
\title{combine.kernels2}
\usage{
combine.kernels2(
  ...,
  scale = TRUE,
  method = c("full-UMKL", "STATIS-UMKL", "sparse-UMKL"),
  knn = 5,
  rho = 20
)
}
\arguments{
\item{method}{methods for integrating multiple kernel matrices,
three methohs to choose: "full-UMKL", "STATIS-UMKL", "sparse-UMKL"}

\item{knn}{= 5 Number of nearest neighbors}

\item{rho}{= 20}

\item{x}{an (n x n) simlarity data matrix of measurements of cancer patients}

\item{y}{an (n x n) simlarity matrix}

\item{z}{an (n x n) simlarity matrix}
}
\value{
similarity matrix of multi-omics
}
\description{
combine.kernels2
}
\examples{
combine.kernels2(x = mRNA_kernel_noc,y = miRNA_kernel_noc,z = methy_kernel_noc,method = "sparse-UMKL", scale=F)

}
